+++ src/main/java/net/enzo/plus/common/container/slot/InfinitySlot.java
@@ -3,6 +3,7 @@ package net.enzo.plus.common.container.slot;
 import net.enzo.plus.common.Config;
 import net.minecraft.inventory.IInventory;
 import net.minecraft.inventory.Slot;
+import net.minecraft.item.ItemStack;
 
 public class InfinitySlot extends Slot {
     public InfinitySlot(IInventory p_i1824_1_, int p_i1824_2_, int p_i1824_3_, int p_i1824_4_) {
@@ -12,5 +13,10 @@ public class InfinitySlot extends Slot {
     @Override
     public int getSlotStackLimit() {
         return !Config.boringChest ? Integer.MAX_VALUE : 64; // Future boring chest config placeholder
+    }
+
+    @Override
+    public void putStack(ItemStack stack) {
+        super.putStack(stack);
     }
 }