+++ src/main/java/net/enzo/plus/common/item/tools/ItemInfinitySword.java
@@ -3,24 +3,37 @@ package net.enzo.plus.common.item.tools;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import net.enzo.plus.PlusMod;
+import net.enzo.plus.common.Config;
 import net.enzo.plus.common.entities.EntityImmortalItem;
 import net.enzo.plus.common.item.InfinityItems;
+import net.enzo.plus.common.item.util.ItemSwordCooler;
 import net.enzo.plus.common.misc.DamageSourceInfinity;
+import net.minecraft.client.renderer.texture.IIconRegister;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.enchantment.Enchantment;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.effect.EntityLightningBolt;
+import net.minecraft.entity.item.EntityXPOrb;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.EnumRarity;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemSword;
+import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EntityDamageSource;
 import net.minecraft.util.EnumChatFormatting;
+import net.minecraft.util.IIcon;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.EnumHelper;
 
 import java.util.List;
+import java.util.Random;
 
-public class ItemInfinitySword extends ItemSword {
+public class ItemInfinitySword extends ItemSwordCooler {
+    public static Random randy = new Random();
+    public IIcon farmer;
     private static final ToolMaterial opSword = EnumHelper.addToolMaterial("INFINITY_SWORD", 32, 9999, 9999F, -3.0F, 32);
     public ItemInfinitySword() {
         super(opSword);
@@ -28,6 +41,49 @@ public class ItemInfinitySword extends ItemSword {
         setTextureName("plus:infinity_sword");
         setCreativeTab(PlusMod.tab);
         setMaxDamage(0);
+    }
+
+    @Override
+    public ItemStack onItemRightClick(ItemStack stack, World world, EntityPlayer player) {
+        if (player.isSneaking()) {
+            NBTTagCompound tags = stack.getTagCompound();
+            if (tags == null) {
+                tags = new NBTTagCompound();
+                stack.setTagCompound(tags);
+            }
+            tags.setBoolean("farm", !tags.getBoolean("farm"));
+            for (int i = 0; i < randy.nextInt(30); i++) {
+                world.spawnEntityInWorld(new EntityLightningBolt(world, player.posX, player.posY, player.posZ));
+            }
+            //player.swingItem();
+        }
+
+        return stack;
+    }
+
+    @Override
+    public void registerIcons(IIconRegister r) {
+        super.registerIcons(r);
+
+        this.itemIcon = r.registerIcon("plus:infinity_sword");
+        farmer = r.registerIcon("plus:infinity_farm");
+    }
+
+    @Override
+    public IIcon getIcon(ItemStack stack, int pass) {
+        NBTTagCompound tags = stack.getTagCompound();
+        if(tags != null){
+            if(tags.getBoolean("farm"))
+                return farmer;
+        }
+        return itemIcon;
+
+    }
+
+    @SideOnly(Side.CLIENT)
+    @Override
+    public IIcon getIconIndex(ItemStack stack) {
+        return getIcon(stack, 0);
     }
 
     @SuppressWarnings({"unchecked", "rawtypes"})
@@ -59,10 +115,31 @@ public class ItemInfinitySword extends ItemSword {
     }
 
     @Override
+    public void getSubItems(Item item, CreativeTabs tabs, List list) {
+        ItemStack sword = new ItemStack(this);
+        if (!Config.boringSword)
+            sword.addEnchantment(Enchantment.looting, 100);
+        else
+            System.out.println("Plus: Shit PC :D");
+        list.add(sword);
+    }
+
+    @Override
     public boolean onLeftClickEntity(ItemStack stack, EntityPlayer player, Entity entity) {
         if(!entity.worldObj.isRemote && entity instanceof EntityPlayer) {
             EntityPlayer victim = (EntityPlayer)entity;
-            if(victim.capabilities.isCreativeMode && !victim.isDead && victim.getHealth() > 0 && !InfinityItems.isPlus(victim)){
+            if(victim.capabilities.isCreativeMode && !victim.isDead && victim.getHealth() > 0 && !InfinityItems.isPlus(victim) && !stack.getTagCompound().getBoolean("farm")){
+                victim.func_110142_aN().func_94547_a(new DamageSourceInfinity(player), victim.getHealth(), victim.getHealth());
+                victim.setHealth(0);
+                victim.onDeath(new EntityDamageSource("infinity", player));
+                //player.addStat(Achievements.creative_kill, 1);
+                return true;
+            } else if (victim.capabilities.isCreativeMode && !victim.isDead && victim.getHealth() > 0 && !InfinityItems.isPlus(victim) && stack.getTagCompound().getBoolean("farm")) {
+                    //player.getEntityWorld().spawnEntityInWorld(new EntityXPOrb(player.getEntityWorld(), player.posX, player.posZ, player.posY, 883));
+                player.addExperienceLevel(696969);
+                for (int i = 0; i < 43; i++) {
+                    entity.worldObj.spawnEntityInWorld(new EntityLightningBolt(entity.worldObj, entity.posX, entity.posY, entity.posZ));
+                }
                 victim.func_110142_aN().func_94547_a(new DamageSourceInfinity(player), victim.getHealth(), victim.getHealth());
                 victim.setHealth(0);
                 victim.onDeath(new EntityDamageSource("infinity", player));