+++ src/main/java/net/enzo/plus/common/item/util/ItemSwordCooler.java
@@ -1,0 +1,10 @@
+package net.enzo.plus.common.item.util;
+
+import net.minecraft.item.ItemSword;
+
+public class ItemSwordCooler extends ItemSword {
+    // SUCK MY DICK WITHERZILA XDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD
+    public ItemSwordCooler(ToolMaterial p_i45356_1_) {
+        super(p_i45356_1_);
+    }
+}